app/requirements.txt
pyyaml
python-dotenv
hvac
